# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fashion API",
    "description": "APIs for fashion service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/v1/brands/{brand_id}": {
      "get": {
        "tags": [
          "Brands"
        ],
        "summary": "특정 브랜드 정보",
        "operationId": "get_1",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "정의되지 않은 브랜드에 대한 정보가 요청됨",
            "content": {
              "application/json": {
                "examples": {
                  "ResourceNotFoundException": {
                    "description": "브랜드 id가 존재하지 않은 경우",
                    "value": {
                      "statusCode": 404,
                      "message": "Brand not found with id: id",
                      "details": "uri=/v1/brands/{brand_id}"
                    }
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BrandDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "summary": "[구현4] 브랜드 수정",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BrandDTO"
                }
              }
            }
          },
          "409": {
            "description": "중복된 이름의 브랜드로 수정하려는 경우",
            "content": {
              "application/json": {
                "examples": {
                  "DuplicateResourceException": {
                    "description": "A 브랜드가 이미 존재하는데 A로 이름을 수정하려는 경우",
                    "value": {
                      "statusCode": 409,
                      "message": "Brand name A already exists",
                      "details": "uri=/v1/brands"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "정의되지 않은 브랜드에 대한 브랜드 수정이 요청됨",
            "content": {
              "application/json": {
                "examples": {
                  "ResourceNotFoundException": {
                    "description": "상품 id가 존재하지 않은 경우",
                    "value": {
                      "statusCode": 404,
                      "message": "Brand not found with id: id",
                      "details": "uri=/v1/brands/{brand_id}"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "summary": "[구현4] 브랜드 삭제",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "정의되지 않은 브랜드에 대한 브랜드 삭제가 요청됨",
            "content": {
              "application/json": {
                "examples": {
                  "ResourceNotFoundException": {
                    "description": "브랜드 id가 존재하지 않은 경우",
                    "value": {
                      "statusCode": 404,
                      "message": "Brand not found with id: id",
                      "details": "uri=/v1/brands/{brand_id}"
                    }
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/brands": {
      "get": {
        "tags": [
          "Brands"
        ],
        "summary": "전체 브랜드 목록",
        "operationId": "list_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BrandListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "summary": "[구현4] 브랜드 추가",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BrandDTO"
                }
              }
            }
          },
          "409": {
            "description": "이미 존재하는 이름의 브랜드를 추가하려 할 때",
            "content": {
              "application/json": {
                "examples": {
                  "DuplicateResourceException": {
                    "description": "A 브랜드가 이미 존재하는데 A브랜드를 추가하려는 경우",
                    "value": {
                      "statusCode": 409,
                      "message": "Brand name A already exists",
                      "details": "uri=/v1/brands"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrandDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "brandName": {
            "type": "string"
          }
        }
      },
...
      "BrandUpdateRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "brandName": {
            "type": "string"
          }
        }
      },
...
      "BrandCreateRequest": {
        "type": "object",
        "properties": {
          "brandName": {
            "type": "string"
          }
        }
      },
...
      "BrandListResponse": {
        "type": "object",
        "properties": {
          "brands": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrandDTO"
            }
          }
        }
      }
    }
  }
}