# src/main/java/com/cy/fashion/brand/dto/request/BrandCreateRequest.java
package com.cy.fashion.brand.dto.request;

import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

@NoArgsConstructor
@AllArgsConstructor
public class BrandCreateRequest {
    public String brandName;
}